Fix create account with avatar

+++ app/Http/Controllers/AdminController.php
@@ -310,7 +310,10 @@ class AdminController extends Controller
             'select_file' => 'required|mimes:xls,xlsx'
         ]);
 
-        User::truncate();
+        if($request->input('select') == 'refesh'){
+            User::truncate();
+        }
+
         Excel::import(new UsersImport, $request->file('select_file'));
         return back()->with('success', 'Excel data imported successfully!');
     }
